+++ src/main/java/woowacourse/shoppingcart/dto/OrderedProductDto.java
@@ -1,6 +1,6 @@
 package woowacourse.shoppingcart.dto;
 
-import woowacourse.shoppingcart.domain.OrderDetail;
+import woowacourse.shoppingcart.domain.order.OrderDetail;
 import woowacourse.shoppingcart.domain.product.Product;
 
 public class OrderedProductDto {
