+++ src/main/java/woowacourse/shoppingcart/service/OrderService.java
@@ -9,14 +9,14 @@ import org.springframework.transaction.annotation.Transactional;
 
 import woowacourse.shoppingcart.dao.CustomerDao;
 import woowacourse.shoppingcart.dao.ProductDao;
-import woowacourse.shoppingcart.domain.OrderDetail;
-import woowacourse.shoppingcart.domain.OrderRepository;
-import woowacourse.shoppingcart.domain.Orders;
 import woowacourse.shoppingcart.domain.Quantity;
 import woowacourse.shoppingcart.domain.cart.CartItem;
 import woowacourse.shoppingcart.domain.cart.CartItemRepository;
 import woowacourse.shoppingcart.domain.cart.CartItems;
 import woowacourse.shoppingcart.domain.customer.Customer;
+import woowacourse.shoppingcart.domain.order.OrderDetail;
+import woowacourse.shoppingcart.domain.order.OrderRepository;
+import woowacourse.shoppingcart.domain.order.Orders;
 import woowacourse.shoppingcart.dto.OrderRequest;
 import woowacourse.shoppingcart.dto.OrderResponse;
